Adding a wrapper to use the gcvspl.f algorythm
Wrapper for the gcvspl.f algorithm was used, in order to be able to use
the gcvspl algorithm for baseline subtraction and data smoothing. The
function linbaseline function in spectratool.py actually points to the
/charleslib/gvspl/gcvspline.py function… Maybe this part has to be
improved. But should work for now. Otherwise you could directly
construct your spline using the gcvspline.py function and avoid using
the linbaseline function framework…