Changes for using gcvspl baselines
Changes have been made in the scripts in order to use the gcvspl
baseline instead of the univariate baseline.

Readme has been a little bit completed.

Spectratools.py has been modified. Now we can input directly x y
spectra with using gcvspl, which assumes in this case that the errors
are sqrt(y).